+++ JudoPayDotNet/Models/WebPaymentRequestModel.cs
@@ -140,6 +140,13 @@ namespace JudoPayDotNet.Models
         [DataMember(EmitDefaultValue = false)]
         public PrimaryAccountDetailsModel PrimaryAccountDetails { get; set; }
 
+        /// <summary>
+        /// Only set this flag on payment sessions for PreAuths, to authenticate a card holder with 3D secure without
+        /// performing payment authorisation.
+        /// </summary>
+        [DataMember(EmitDefaultValue = false)]
+        public bool? DelayedAuthorisation { get; set; }
+
         private Dictionary<string, string> _httpHeaders;
 
         /// <summary>